Update patch set 3
Patch Set 3:

> > Have you flashed the latest firmware? But it should boot okay
 > anyway.
 > 
 > No I didn't that time. There's no firmware listed on the LineageOS
 > wiki page, just some random files on XDA... Why not include that in
 > the ROM or simply provide valid links to the required firmware?
 > 
 > Alright, what have I tried so far:
 > * CM14.1-firmware_20161205 by Psy_Man - still the same result,
 > doesn't boot
 > * CM14.1_Firmware_kenzo_12.04.2017 by TheStrix - locked my
 > bootloader (WTF), obviously doesn't let the phone to boot. I Will
 > unlock it later and see what happens.

Wohoo, after unlocking the bootloader it immediately worked after a reboot! It would be really-really nice to get the proper firmware info (preferably the one that doesn't mess with the booloader) on the wiki page.

Thank you guys for implementing the goodix support and hinting about hte firmware!

Patch-set: 3